Moved the WorkStatusCategory and BacklogCategory enums into Moda.Common.Domain\Enums\Work.
