+++ Moda.Common/src/Moda.Common.Domain/Enums/Work/BacklogCategory.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace Moda.Common.Domain.Enums;
+namespace Moda.Common.Domain.Enums.Work;
 public enum BacklogCategory
 {
     [Display(Name = "Portfolio", Description = "Portfolio backlogs provide a way to group related items into a hierarchical structure.  This is the only backlog category that allows multiple backlog levels.", Order = 1)]
