+++ Moda.Integrations/src/Moda.Integrations.AzureDevOps/Models/Contracts/AzdoBacklogLevel.cs
@@ -1,5 +1,5 @@
 ﻿using Moda.Common.Application.Interfaces.ExternalWork;
-using Moda.Common.Domain.Enums;
+using Moda.Common.Domain.Enums.Work;
 
 namespace Moda.Integrations.AzureDevOps.Models.Contracts;
 public sealed record AzdoBacklogLevel : IExternalBacklogLevel
