+++ Moda.Services/Moda.Work/src/Moda.Work.Application/BacklogLevels/Dtos/BacklogLevelDto.cs
@@ -1,4 +1,4 @@
-﻿using Moda.Common.Domain.Enums;
+﻿using Moda.Common.Domain.Enums.Work;
 
 namespace Moda.Work.Application.BacklogLevels.Dtos;
 public sealed record BacklogLevelDto : IMapFrom<BacklogLevel>
