+++ Moda.Services/Moda.Work/src/Moda.Work.Application/WorkStatusCategories/Queries/GetWorkStatusCategoriesQuery.cs
@@ -1,4 +1,6 @@
-﻿namespace Moda.Work.Application.WorkStatusCategories.Queries;
+﻿using Moda.Common.Domain.Enums.Work;
+
+namespace Moda.Work.Application.WorkStatusCategories.Queries;
 
 public sealed record GetWorkStatusCategoriesQuery : IQuery<IReadOnlyList<WorkStatusCategoryListDto>> { }
