+++ Moda.Services/Moda.Work/src/Moda.Work.Domain/Models/BacklogLevel.cs
@@ -1,5 +1,6 @@
 ﻿using Ardalis.GuardClauses;
 using CSharpFunctionalExtensions;
+using Moda.Common.Domain.Enums.Work;
 using Moda.Common.Extensions;
 using NodaTime;
