+++ Moda.Services/Moda.Work/src/Moda.Work.Domain/Models/BacklogLevelScheme.cs
@@ -1,5 +1,6 @@
 ﻿using Ardalis.GuardClauses;
 using CSharpFunctionalExtensions;
+using Moda.Common.Domain.Enums.Work;
 using NodaTime;
 
 namespace Moda.Work.Domain.Models;
